pyproject.toml
[project]
name = "CBattle"
version = "0.0.1a"
description = "Interactive battle package for the Ballsdex Discord bot."
authors = [
    'Cayla "DotZZ" Selby <official.dotzz@gmail.com>',
]
license = "MIT"
requires-python = ">=3.13"
dependencies = [
    "discord.py==2.5.0"
]

[dependency-groups]
dev = [
    "ruff>=0.12.9"
]

[tool.ruff]
line-length = 120
indent-width = 4
target-version = "py313"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = true
line-ending = "lf"
docstring-code-format = true

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E203"]

[tool.ruff.lint.isort]
known-first-party = ["ballsdex"]
split-on-trailing-comma = false

[tool.ruff.lint.per-file-ignores]
"CBattle/github/*.py" = ["F704", "F706", "F821"]
